+++ Artsee-Backend/src/test/java/com/artsee/backend/dao/TestArtseePersistance.java
@@ -228,7 +228,7 @@ public class TestArtseePersistance {
 		Integer artworkID = 122;
 		String name = "ArtworkTestName";
 		String description = "Artwork description test";
-		Float price = 1500.f;
+		Integer price = 1500;
 		Date dateOfCreation = java.sql.Date.valueOf(LocalDate.of(2020, Month.SEPTEMBER, 15));
 		Integer numInStock = 3;
 		Artwork artwork = createTestArtworkWithoutArtist(artworkID , name, description, price, dateOfCreation, numInStock);
@@ -413,7 +413,7 @@ public class TestArtseePersistance {
 		return artist;
 	}
 	
-	public Artwork createTestArtworkWithoutArtist(Integer artworkID , String name, String description, Float price, Date dateOfCreation , Integer numInStock){		
+	public Artwork createTestArtworkWithoutArtist(Integer artworkID , String name, String description, Integer price, Date dateOfCreation , Integer numInStock){		
 		Artwork artwork = new Artwork();
 		artwork.setArtworkID(artworkID);
 		artwork.setName(name);
